add more pinger tests
