+++ test/net/azib/ipscan/core/net/CombinedUnprivilegedPingerTest.java
@@ -1,0 +1,7 @@
+package net.azib.ipscan.core.net;
+
+public class CombinedUnprivilegedPingerTest extends AbstractPingerTest {
+	public CombinedUnprivilegedPingerTest() {
+		super(new CombinedUnprivilegedPinger(10));
+	}
+}
