+++ test/net/azib/ipscan/core/net/UDPPingerTest.java
@@ -1,0 +1,7 @@
+package net.azib.ipscan.core.net;
+
+public class UDPPingerTest extends AbstractPingerTest {
+	public UDPPingerTest() {
+		super(new UDPPinger(10));
+	}
+}
